+++ CHANGELOG.md
@@ -13,9 +13,11 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 
 - support for matching line number in `abort-on-stack-trace` scriptlet
   when `inlineScript` or `injectedScript` option is used [#439]
+- new values to `set-cookie` and `set-cookie-reload` scriptlets: `checked`, `unchecked` [#444]
 
 [Unreleased]: https://github.com/AdguardTeam/Scriptlets/compare/v1.11.16...HEAD
 [#439]: https://github.com/AdguardTeam/Scriptlets/issues/439
+[#444]: https://github.com/AdguardTeam/Scriptlets/issues/444
 
 ## [v1.11.16] - 2024-08-01